MDT-89: chore - rename medicine to medication
