+++ medit-frontend/src/components/home/AddMedication/AddImageProfile.tsx
