+++ medit-frontend/src/components/home/AddMedication/AddInfo.tsx
