+++ medit-frontend/src/components/home/home.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { Navbar } from "../Navbar/Navbar.tsx";
-import { MedicineComponent } from "./MedicineComponent/MedicineComponent.tsx";
+import { MedicationComponent } from "./MedicationComponent/MedicationComponent.tsx";
 import { Calendar } from "./calendar/calendar.tsx";
 import "../../index.css";
 import { FilterButton } from "./FilterButton/FilterButton.tsx";
@@ -49,7 +49,7 @@ export const Home: React.FC = () => {
       <Calendar onDateChange={handleDateChange} />
       <FilterButton />
       {medications.map((med) => (
-        <MedicineComponent key={med.id} medication={med} />
+        <MedicationComponent key={med.id} medication={med} />
       ))}
       <ReminderModal
         showSetReminder={showSetReminder}
